# frontend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusEats v1",
    "description": "单校区点餐演示；所有支付退款均为模拟",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/admin/shop": {
      "get": {
        "tags": [
          "shop-controller"
        ],
        "operationId": "shop",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ShopView"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "shop-controller"
        ],
        "operationId": "configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shop"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/quotas": {
      "get": {
        "tags": [
          "shop-controller"
        ],
        "operationId": "quotas",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "shop-controller"
        ],
        "operationId": "quota",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Quota"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/buildings/{id}": {
      "put": {
        "tags": [
          "shop-controller"
        ],
        "operationId": "buildingUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Building"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/buildings": {
      "get": {
        "tags": [
          "shop-controller"
        ],
        "operationId": "adminBuildings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BuildingView"
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "shop-controller"
        ],
        "operationId": "building",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Building"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Shop": {
        "type": "object",
        "properties": {
          "open": {
            "type": "boolean"
          },
          "deliveryFee": {
            "type": "string",
            "example": "12.00"
          },
          "packagingFee": {
            "type": "string",
            "example": "12.00"
          },
          "hours": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "([01][0-9]|2[0-3]):[0-5][0-9]-(?:([01][0-9]|2[0-3]):[0-5][0-9]|24:00)"
            },
            "maxItems": 8,
            "minItems": 0
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "minLength": 0
          }
        },
        "required": [
          "deliveryFee",
          "hours",
          "packagingFee"
        ]
      },
      "Quota": {
        "type": "object",
        "properties": {
          "dishId": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "maximum": 100000,
            "minimum": 0
          },
          "defaultQuota": {
            "type": "boolean"
          }
        },
        "required": [
          "dishId"
        ]
      },
...
      "Building": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "BuildingView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
      "ShopView": {
        "type": "object",
        "properties": {
          "open": {
            "type": "boolean"
          },
          "accepting": {
            "type": "boolean"
          },
          "deliveryFee": {
            "type": "string",
            "example": "12.00"
          },
          "packagingFee": {
            "type": "string",
            "example": "12.00"
          },
          "hours": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "phone": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}